refactor: litellm noise suppressions
