+++ contextgem/internal/utils.py
@@ -24,17 +24,13 @@ from __future__ import annotations
 
 import asyncio
 import concurrent.futures
-import functools
-import inspect
 import json
 import re
-import warnings
 from collections import defaultdict
-from collections.abc import Callable, Coroutine, Generator
-from contextlib import contextmanager
+from collections.abc import Callable, Coroutine
 from importlib import resources
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Literal, TypeVar, cast, get_args
+from typing import TYPE_CHECKING, Any, Literal, TypeVar, get_args
 
 from jinja2 import Environment, Template, nodes
 from wtpsplit_lite import SaT
@@ -47,6 +43,7 @@ from contextgem.internal.llm_output_structs.concept_structs import (
     _get_concept_extraction_output_struct,
 )
 from contextgem.internal.loggers import logger
+from contextgem.internal.suppressions import _suppress_litellm_warnings
 from contextgem.internal.typings.aliases import (
     AsyncCalsAndKwargs,
     ExtractedInstanceType,
@@ -64,7 +61,6 @@ if TYPE_CHECKING:
 
 
 T = TypeVar("T")
-F = TypeVar("F", bound=Callable[..., Any])
 
 
 def _get_template(
@@ -467,110 +463,6 @@ async def _async_multi_executor(
     await asyncio.gather(*tasks)
     logger.debug(f"Completed all tasks ({len(tasks)})")
     return results
-
-
-@contextmanager
-def _suppress_litellm_warnings_context() -> Generator[None, None, None]:
-    """
-    Context manager that suppresses Pydantic and httpx deprecation and serialization warnings.
-
-    This context manager temporarily suppresses DeprecationWarnings and specific UserWarnings
-    from the Pydantic and httpx modules. These warnings are typically generated by litellm's
-    internal usage of these libraries and are not actionable by users.
-
-    Specifically suppresses:
-    - Pydantic DeprecationWarnings
-    - Pydantic serialization UserWarnings
-    - httpx DeprecationWarnings about content parameter usage (typically when using local LLMs)
-
-    The context manager uses Python's warnings.catch_warnings() to create a temporary
-    warning filter that ignores the specified warnings only within its scope.
-
-    :return: A generator that yields None and expects no values to be sent back.
-    :rtype: Generator[None, None, None]
-    """
-    with warnings.catch_warnings():
-        # Pydantic
-        warnings.filterwarnings(
-            "ignore", category=DeprecationWarning, module="pydantic"
-        )
-        warnings.filterwarnings(
-            "ignore",
-            category=UserWarning,
-            module="pydantic",
-            message=r"^Pydantic serializer warnings",
-        )
-        # httpx (typically when using local LLMs)
-        warnings.filterwarnings(
-            "ignore",
-            category=DeprecationWarning,
-            module="httpx",
-            message="Use 'content=<...>'",
-        )
-        yield
-
-
-def _suppress_litellm_warnings(func: F) -> F:
-    """
-    A decorator that suppresses warnings related to Pydantic and httpx deprecation
-    and serialization in litellm>1.71.1 (latest available version as of 2025-07-10).
-
-    This decorator wraps both synchronous and asynchronous functions to suppress
-    Pydantic and httpx warnings that originate from litellm's internal usage.
-    These warnings are not actionable by users and should not be displayed.
-
-    Suppression rationale:
-    - The warnings originate from litellm's internal pydantic and httpx usage, not user code
-    - These are dependency warnings that users cannot fix and should not see
-
-    :param func: The function to wrap with warning suppression. Can be either
-        synchronous or asynchronous.
-    :type func: F
-    :return: The wrapped function with the same signature as the input function,
-        but with Pydantic and httpx deprecation and serialization warnings suppressed
-        during execution.
-    :rtype: F
-
-    TODO: Remove deprecation-related suppression when deprecation warnings are fixed
-    in a future litellm release.
-
-    TODO: Remove serialization-related suppression when the related issue is fixed
-    in a future litellm release: https://github.com/BerriAI/litellm/issues/11759
-    """
-
-    @functools.wraps(func)
-    async def _async_wrapper(*args, **kwargs) -> Any:
-        """
-        Async wrapper that suppresses Pydantic and httpx deprecation and serialization
-        warnings during async function execution.
-
-        :param args: Positional arguments to pass to the wrapped function.
-        :param kwargs: Keyword arguments to pass to the wrapped function.
-        :return: The result of the wrapped async function.
-        :rtype: Any
-        """
-        with _suppress_litellm_warnings_context():
-            return await func(*args, **kwargs)
-
-    @functools.wraps(func)
-    def _sync_wrapper(*args, **kwargs) -> Any:
-        """
-        Sync wrapper that suppresses Pydantic and httpx deprecation and serialization
-        warnings during sync function execution.
-
-        :param args: Positional arguments to pass to the wrapped function.
-        :param kwargs: Keyword arguments to pass to the wrapped function.
-        :return: The result of the wrapped sync function.
-        :rtype: Any
-        """
-        with _suppress_litellm_warnings_context():
-            return func(*args, **kwargs)
-
-    # Safe cast: tell type checker that functools.wraps preserves the original function's type.
-    # The wrapper functions maintain the same signature as the original function.
-    return cast(
-        F, _async_wrapper if inspect.iscoroutinefunction(func) else _sync_wrapper
-    )
 
 
 @_suppress_litellm_warnings
